adding pix key limit validation, response json message and some refactoring

+++ src/main/java/com/pix/api/dto/CreatePixKeyDTO.java
@@ -2,15 +2,16 @@ package com.pix.api.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreatePixKeyDTO {
 
-
     private String keyType;
     private String keyValue;
-    private int agencyNumber;
+    private BigDecimal agencyNumber;
     private String accountType;
-    private int accountNumber;
+    private BigDecimal accountNumber;
     private String accountHolderName;
     private String accountHolderLastName;
     private String personType;
